+++ modules/core/errors/stacktrace-error.ts
