--- src/index.ts
@@ -1,3 +1,0 @@
-export { RunnerResolver } from '@modules/promise/runner.resolver';
-export { IRunnerError } from './commands/runner-error';
-export { RunnerErrorCode, RunnerErrorMessages } from './errors/runners-errors';
