+++ test/collection/execute.ts
@@ -1,4 +1,5 @@
-import { IRunnerError, RunnerErrorCode, RunnerErrorMessages } from "../../src";
+import { IRunnerError } from "@core/commands/runner-error";
+import { RunnerErrorCode, RunnerErrorMessages } from "@core/errors/runners-errors";
 import { resolver } from "../common";
 import { CalcAmountRunner } from "../common/calc-amount.runner";
 import { DelayRunner } from "../common/delay.runner";
